Implementing RETAKE, ABORT and WARN + avoiding stale UDP messages
Now properly check if ball is touched during a game interruption and performs
the appropriate action. To obtain a reliable mechanism without introducing
new memory variables, I had to fix the issue of stale UDP messages described
below.

`game_controller_receive()` only read the most recent UDP message, one of the
impact of that choice is that when simulator was waiting to sync with
GameController, it had to tick several time to get rid of all the stale
messages. This has been modified to ensure that the function now read all the
messages and only takes into account the most recent one.