Forceful contact: testing competition for the ball [8/9]
This scenario tests the behaviors of the AutoRef when two robots collide while
they are competing for the ball.

One of the test does not pass because after being penalized, the robot still has
inertia and falls inside the field.

A small change was added to the referee to distinguish the messages between
'charge' and 'charge from behind'.