RoboCup: bandwidth limitation system changed
According to the comments of the teams, having an individual limit per robot
rather than a limit per team would not have significant impact.

This commit makes implementation of the bandwidth limitation system much simpler
and solves an existing issue where the size of the answer was not properly
changed.

The Player is now also made aware of the number of robots in the team.