Ball positioning at the beginning of second half with testing
Fixes #216

From logs of games, it seems that the what was expected to be dead code was
still used with the following side effect: ball was not respawned at an
appropriate location.

Additionnally, the ball is now moved away from the field before kick-off.

The test for `second half` now tests for ball position as well.